feat(DateService): migrated getChartData()
