+++ src/app/models/chart-data.model.ts
@@ -1,0 +1,7 @@
+import { Column, Row } from 'angular-google-charts';
+
+export interface ChartData {
+	columns: Column[];
+	rows: Row[];
+	options: object;
+}
