+++ src/app/models/index.ts
@@ -1,3 +1,4 @@
+export * from './chart-data.model';
 export * from './github-repository-contributors.model';
 export * from './github-repository-languages.model';
 export * from './github-repository.model';
