+++ src/app/services/date/date.service.ts
@@ -1,10 +1,42 @@
-import { Injectable } from '@angular/core';
+import { formatNumber } from '@angular/common';
+import { inject, Injectable, LOCALE_ID } from '@angular/core';
+import { Column } from 'angular-google-charts';
 import { addDays, differenceInDays, format, isEqual } from 'date-fns';
+import { ChartData, RegistryData } from '../../models';
 
 @Injectable({
 	providedIn: 'root',
 })
 export class DateService {
+	private readonly locale = inject(LOCALE_ID);
+
+	getChartData(apiDatas: RegistryData[], start: Date, end: Date): ChartData {
+		const columns: Column[] = [
+			{ type: 'string', label: 'Date' },
+			...apiDatas.map(({ packageName, total }) => ({
+				type: 'number',
+				label: `${packageName} (${formatNumber(total, this.locale)})`,
+			})),
+		];
+
+		const dates = this.getDateRange(start, end);
+
+		// TODO: get formatter given range
+		// MM/dd vs MM/dd/yy
+		const rows = dates.map((date, i) => {
+			return [format(date, 'MM/dd/yy'), ...apiDatas.map((apiData) => apiData.range[i])];
+		});
+
+		const options = {
+			chart: {
+				title: 'Downloads',
+				subtitle: 'per day for a given period of specific package(s)',
+			},
+			height: 400,
+		};
+
+		return { columns, rows, options };
+	}
 	getFormattedDateString(date: Date): string {
 		return format(date, 'yyyy-MM-dd');
 	}
